refactor: extract Sidebar component from App.tsx
